🚚 Move: `Achievement` struct to `achievement` module [#63]
